+++ hw09_struct_validator/.golangci.yml
@@ -1,0 +1,3 @@
+linters:
+  enable:
+    - copyloopvar
