+++ hw09_struct_validator/validator.go
@@ -4,8 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"regexp"
-	"strconv"
 	"strings"
 )
 
@@ -29,7 +27,7 @@ func Validate(v interface{}) error {
 	var errorsResult ValidationErrors
 
 	rv := reflect.ValueOf(v)
-	if rv.Kind() != reflect.Ptr {
+	if rv.Kind() == reflect.Ptr {
 		if rv.IsNil() {
 			return errors.New("pointer is nil")
 		}
@@ -88,9 +86,24 @@ func Validate(v interface{}) error {
 		}
 
 		switch valField.Kind() {
+
+		case reflect.String:
+			validators, err := CreateStrRuleValidators(tag)
+			if err != nil {
+				return fmt.Errorf("%s: %w", nameField, err)
+			}
+			s := valField.String() //valField.String()
+			for _, valid := range validators {
+				if err := valid.Validate(s); err != nil {
+					errorsResult = append(errorsResult, ValidationError{
+						Field: fmt.Sprintf("%s.%s", field.Name, nameField),
+						Err:   err,
+					})
+				}
+			}
 		case reflect.Struct:
 			// парсинг
-			validators, err := createStrRuleValidators(tag)
+			validators, err := CreateStrRuleValidators(tag)
 			if err != nil {
 				return fmt.Errorf("%s: %w", nameField, err)
 			}
@@ -113,11 +126,11 @@ func Validate(v interface{}) error {
 				//case reflect.Struct:
 				case reflect.String:
 					// todo: в отдельную - копипаст пошел
-					validators, err := createStrRuleValidators(tag)
+					validators, err := CreateStrRuleValidators(tag)
 					if err != nil {
 						return fmt.Errorf("%s: %w", itemName, err)
 					}
-					s := valField.String()
+					s := item.String() //valField.String()
 					for _, valid := range validators {
 						if err := valid.Validate(s); err != nil {
 							errorsResult = append(errorsResult, ValidationError{
@@ -127,7 +140,7 @@ func Validate(v interface{}) error {
 						}
 					}
 				case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-					validators, err := createIntRuleValidators(tag)
+					validators, err := CreateIntRuleValidators(tag)
 					if err != nil {
 						return fmt.Errorf("%s: %w", itemName, err)
 					}
@@ -147,7 +160,7 @@ func Validate(v interface{}) error {
 			}
 
 		case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-			validators, err := createIntRuleValidators(tag)
+			validators, err := CreateIntRuleValidators(tag)
 			if err != nil {
 				return fmt.Errorf("%s: %w", nameField, err)
 			}
@@ -179,168 +192,4 @@ func Validate(v interface{}) error {
 
 type Validator[T any] interface {
 	Validate(T) error
-}
-
-type funValidator[T any] func(T) error //result,error
-
-func (f funValidator[T]) Validate(v T) error {
-	return f(v)
-}
-
-func lenValidator(data int) Validator[string] {
-	return funValidator[string](func(s string) error {
-		if len(s) != data {
-			//return errors.New("len fail")
-			return fmt.Errorf("length of %s is %d, want %d", s, len(s), data)
-		}
-		return nil
-	})
-}
-
-func regValidator(pat string) Validator[string] {
-	r, err := regexp.Compile(pat)
-	if err != nil {
-		return funValidator[string](func(s string) error {
-			return fmt.Errorf("invalid regular expression: %s - %w", s, err)
-		})
-	}
-	return funValidator[string](func(s string) error {
-
-		if !r.MatchString(s) {
-			// с кавычками пробуем формат %q
-			return fmt.Errorf("invalid regular expression: %q  - %w - %q", s, err, pat)
-		}
-
-		return nil
-	})
-}
-
-func containStrValidator(values []string) Validator[string] {
-	return funValidator[string](func(s string) error {
-		for _, v := range values {
-			if v == s {
-				return nil
-			}
-		}
-		return fmt.Errorf("%q not in %v", s, values)
-	})
-}
-
-func minValidator(min int) Validator[int] {
-	return funValidator[int](func(v int) error {
-		if v < min {
-			return fmt.Errorf("value %d is less than minimum %d", v, min)
-		}
-		return nil
-	})
-}
-
-func maxValidator(max int) Validator[int] {
-	return funValidator[int](func(v int) error {
-		if v > max {
-			return fmt.Errorf("value %d is greater than maximum %d", v, max)
-		}
-		return nil
-	})
-}
-
-func containInValidator(values []int) Validator[int] {
-	return funValidator[int](func(val int) error {
-		for _, v := range values {
-			if v == val {
-				return nil
-			}
-		}
-		return fmt.Errorf("value %d not in %v", val, values)
-	})
-}
-
-func createStrRuleValidators(tag string) ([]Validator[string], error) {
-	var validators []Validator[string]
-	rules := strings.Split(tag, "|")
-	for _, rule := range rules {
-		rule = strings.TrimSpace(rule)
-		if rule == "" {
-			continue
-		}
-		parts := strings.SplitN(rule, ":", 2)
-		if len(parts) != 2 {
-			return nil, fmt.Errorf("invalid tag rule: %q", rule)
-		}
-
-		switch parts[0] {
-		case "len":
-			expected, err := strconv.Atoi(parts[1])
-			if err != nil {
-				return nil, fmt.Errorf("invalid tag rule: %q", rule)
-			}
-			validators = append(validators, lenValidator(expected))
-		case "regexp":
-			validators = append(validators, regValidator(parts[1]))
-
-		case "in":
-			options := strings.Split(parts[1], ",")
-			var items = make([]string, len(options))
-			for _, option := range options {
-				//temp, err := strconv.Atoi(option)
-				//
-				//if err != nil {
-				//	return nil, fmt.Errorf("invalid tag rule: %q - %w ", rule, err)
-				//}
-				items = append(items, strings.TrimSpace(option))
-			}
-			validators = append(validators, containStrValidator(items))
-		default:
-			return nil, fmt.Errorf("invalid tag rule: %q", rule)
-		}
-	}
-	return validators, nil
-}
-
-func createIntRuleValidators(tag string) ([]Validator[int], error) {
-	var validators []Validator[int]
-
-	rules := strings.Split(tag, "|")
-
-	for _, rule := range rules {
-		rule = strings.TrimSpace(rule)
-		if rule == "" {
-			continue
-		}
-		parts := strings.SplitN(rule, ":", 2)
-		if len(parts) != 2 {
-			return nil, fmt.Errorf("invalid tag rule: %q", rule)
-		}
-
-		switch parts[0] {
-		case "min":
-			val, err := strconv.Atoi(parts[1])
-			if err != nil {
-				return nil, fmt.Errorf("invalid tag rule: %q", rule)
-			}
-			validators = append(validators, minValidator(val))
-		case "max":
-			val, err := strconv.Atoi(parts[1])
-			if err != nil {
-				return nil, fmt.Errorf("invalid tag rule: %q", rule)
-			}
-			validators = append(validators, minValidator(val))
-		case "in":
-			options := strings.Split(parts[1], ",")
-			items := make([]int, len(options))
-			for _, option := range options {
-				s := strings.TrimSpace(option)
-				val, err := strconv.Atoi(s)
-				if err != nil {
-					return nil, fmt.Errorf("invalid tag rule: %q - %w", rule, err)
-				}
-				items = append(items, val)
-			}
-			validators = append(validators, containInValidator(items))
-		default:
-			return nil, fmt.Errorf("invalid tag rule: %q", rule)
-		}
-	}
-
-	return validators, nil
 }
